feat: more server fetches
