+++ web-components/src/components/cards/CreditClassCard/CreditClassCard.tsx
@@ -91,6 +91,7 @@ const CreditClassCard = ({
           <Subtitle
             size="xs"
             color="info.main"
+            component="div"
             sx={{
               fontWeight: 800,
               mb: 3,
@@ -111,13 +112,14 @@ const CreditClassCard = ({
             {parseText(title)}
           </Title>
           <Body
+            component="div"
             size="sm"
             mobileSize="xs"
             sx={{
               mb: 5,
               display: '-webkit-box',
-              '-webkit-line-clamp': '3',
-              '-webkit-box-orient': 'vertical',
+              WebkitLineClamp: '3',
+              WebkitBoxOrient: 'vertical',
               overflow: 'hidden',
             }}
           >
