+++ web-components/src/components/cards/MediaCard/MediaCard.tsx
@@ -147,8 +147,8 @@ export default function MediaCard({
               overflow: 'hidden',
               textOverflow: 'ellipsis',
               display: '-webkit-box',
-              '-webkit-line-clamp': '2',
-              '-webkit-box-orient': 'vertical',
+              WebkitLignClamp: '2',
+              WebkitBoxOrient: 'vertical',
             },
           ]}
         >
