+++ web-marketplace/next.config.mjs
@@ -50,6 +50,21 @@ const nextConfig = {
     position: 'bottom-right',
   },
   // Production environment
+  images: {
+    remotePatterns: [
+      {
+        protocol: 'https',
+        hostname: 'cdn.sanity.io',
+        port: '',
+        pathname: '/images/**',
+      },
+      {
+        protocol: 'https',
+        hostname: 'regen-registry.s3.amazonaws.com',
+        port: '',
+      },
+    ],
+  },
   webpack: (config, { buildId, dev, isServer, defaultLoaders, webpack }) => {
     // https://github.com/sindresorhus/got/issues/345
     config.plugins.push(
