+++ web-marketplace/src/ledger.tsx
@@ -12,6 +12,7 @@ import {
   regenAminoConverters,
   regenProtoRegistry,
 } from '@regen-network/api';
+import { getRPCQueryClient } from 'app/makeRPCQueryClient';
 import { chains } from 'chain-registry';
 
 import { UseStateSetter } from 'types/react/use-state';
@@ -83,20 +84,6 @@ export async function setupSigningClient(
       setLoading(false);
     }
   }
-}
-
-export async function getRPCQueryClient() {
-  const { createRPCQueryClient } = regen.ClientFactory;
-  const client = await createRPCQueryClient({
-    rpcEndpoint: ledgerRPCUri,
-  });
-
-  const { createRPCQueryClient: createRPCQueryIBClient } = ibc.ClientFactory;
-  const ibcClient = await createRPCQueryIBClient({
-    rpcEndpoint: ledgerRPCUri,
-  });
-
-  return { ...ibcClient, ...client };
 }
 
 async function setupRPCQueryClient(
