+++ web-marketplace/src/legacy-pages/Projects/AllProjects/AllProjects.loader.ts
@@ -1,6 +1,6 @@
 import { QueryClient } from '@tanstack/react-query';
+import { getRPCQueryClient } from 'app/makeRPCQueryClient';
 
-import { getRPCQueryClient } from 'ledger';
 import { defaultLocale } from 'lib/i18n/i18n';
 import { getSimplePriceQuery } from 'lib/queries/react-query/coingecko/simplePrice/simplePriceQuery';
 import { getProjectsQuery } from 'lib/queries/react-query/ecocredit/getProjectsQuery/getProjectsQuery';
